Augment preact component definitions with routable props
Allows you to use any Component, FunctionalComponent (if explicitly
typed) or IntrinsicElement with routeable properties (path, default) in
a typesafe way. Meaning no typescript errors and editor code completion.

e.g. previously where you would have to do

```typescript
<Router>
    <Route default component={Home} />
    <Route path="/about" component={About} />
</Router>
```

It now becomes possible to directly use components directly as you would
in a plain javascript project

```typescript
<Router>
    <Home default />
    <About path="/about" />
</Router>
```

Also removes the need for the copy-typescript-definition build step and
dependency
