--- src/match.d.ts
@@ -1,15 +1,0 @@
-import * as preact from 'preact';
-
-import { Link as StaticLink, RoutableProps } from './';
-
-export class Match extends preact.Component<RoutableProps, {}> {
-    render(): preact.VNode;
-}
-
-export interface LinkProps extends JSX.HTMLAttributes {
-    activeClassName: string;
-}
-
-export function Link(props: LinkProps): preact.VNode;
-
-export default Match;
